@@ -1,26 +1,23 @@
 {
   "name": "client",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
-    "@react-pdf/renderer": "^1.6.8",
-    "crypto-random-string": "^3.0.1",
-    "jquery": "^3.4.1",
-    "papaparse": "^4.6.3",
-    "rc-datepicker": "^5.0.15",
-    "react": "^16.12.0",
-    "react-cookies": "^0.1.1",
-    "react-dom": "^16.12.0",
-    "react-pdf": "^4.1.0",
-    "react-request": "^3.1.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "2.1.8",
-    "react-simple-invoice": "^1.0.7",
-    "react-to-print": "^2.5.0",
-    "semantic-ui-calendar-react": "^0.15.3",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.86.0",
-    "semantic-ui-react-input-file": "^1.0.3"
+    "@react-pdf/renderer": "^3.4.2",
+    "jquery": "^3.7.1",
+    "js-cookie": "^3.0.5",
+    "papaparse": "^5.4.1",
+    "pdfjs-dist": "^5.2.133",
+    "react": "^18.2.0",
+    "react-datepicker": "^4.8.0",
+    "react-dom": "^18.2.0",
+    "react-pdf": "^7.6.0",
+    "react-router-dom": "^6.23.0",
+    "react-scripts": "^5.0.1",
+    "react-simple-invoice": "^1.0.8",
+    "react-to-print": "^2.15.0",
+    "semantic-ui-css": "^2.5.0",
+    "semantic-ui-react": "^2.1.5"
   },
   "scripts": {
     "start": "react-scripts start",
