@@ -16,25 +16,25 @@
     "compression": "^1.0.3",
     "cors": "^2.5.2",
     "crypto": "^1.0.1",
-    "db": "^1.1.1",
-    "fast-csv": "^2.4.1",
+    "db": "^6.0.3",
+    "fast-csv": "^5.0.2",
     "helmet": "^3.10.0",
     "loopback": "^3.22.0",
     "loopback-boot": "^2.6.5",
-    "loopback-component-explorer": "^6.2.0",
+    "loopback-component-explorer": "^2.7.0",
     "loopback-connector-mongodb": "^3.9.2",
     "loopback-connector-mysql": "^5.3.1",
-    "loopback-connector-rest": "^3.4.1",
+    "loopback-connector-rest": "^5.0.6",
     "multer": "^1.4.1",
     "mysql": "^2.16.0",
-    "request-promise": "^4.2.4",
+    "request-promise": "^0.0.1",
     "serve-favicon": "^2.0.1",
-    "solr-client": "^0.7.0",
+    "solr-client": "^0.10.0-rc10",
     "strong-error-handler": "^3.0.0",
     "underscore": "^1.9.1"
   },
   "devDependencies": {
-    "eslint": "^3.17.1",
+    "eslint": "^9.28.0",
     "eslint-config-loopback": "^8.0.0"
   },
   "repository": {
