@@ -1,8 +1,9 @@
 import React, {Component} from 'react'
 import * as Constants from '../common/constants'
 import Layout from '../components/Layout'
-import cookie from 'react-cookies'
+import Cookies from 'js-cookie'
 import { Container, Segment, Header, Form, Button, Input, Message, Dimmer, Loader } from 'semantic-ui-react'
+import { withRouter } from '../utils/withRouter';
 
 class AddAmount extends Component {
     state = {
@@ -17,11 +18,11 @@ class AddAmount extends Component {
     }
 
     componentDidMount() {
-        if ( ! cookie.load("userId")) {
-            this.props.history.push("/login")
+        if ( ! Cookies.get("userId")) {
+            this.props.navigate("/login")
         }
         
-        this.setState({userid: cookie.load("userId")})
+        this.setState({userid: Cookies.get("userId")})
     }
 
 
@@ -111,4 +112,4 @@ class AddAmount extends Component {
     }
 }
 
-export default AddAmount
+export default withRouter(AddAmount)
